requirements.txt
alembic==1.17.2
annotated-types==0.7.0
arrow==1.4.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colour==0.1.5
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Admin==2.0.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
mailerlite==0.1.10
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
psycopg==3.2.12
psycopg2==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
spectree==1.5.8
SQLAlchemy==2.0.44
sqlalchemy-citext==1.8.0
SQLAlchemy-Utils==0.42.1
tablib==3.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
